Fix the API tests
