+++ npm/test/dsync/data/groups.ts
@@ -4,6 +4,11 @@ const groups = [
     displayName: 'Developers',
     members: [],
   },
+  {
+    schemas: ['urn:ietf:params:scim:schemas:core:2.0:Group'],
+    displayName: 'Designers',
+    members: [],
+  },
 ];
 
 export default groups;
